Part of WOR-403: split watcher_helpers.py into watcher_helpers.py + watcher_log_parsing.py
