+++ tests/test_watcher_helpers.py
@@ -2,13 +2,9 @@
 
 from __future__ import annotations
 
-import json
 from pathlib import Path
 
 from app.core.watcher.watcher_helpers import (
-    _parse_worker_api_retries,
-    _parse_worker_subagent_spawns,
-    _parse_worker_usage,
     build_worker_cmd,
     build_worker_env,
     check_allowed_paths_overlap,
@@ -257,740 +253,6 @@ def test_default_defers_to_manifest() -> None:
 
 def test_default_hybrid_becomes_cloud() -> None:
     assert resolve_effective_mode("default", "hybrid") == "cloud"
-
-
-# ---------------------------------------------------------------------------
-# _parse_worker_usage
-# ---------------------------------------------------------------------------
-
-
-def _write_log(tmp_path: Path, lines: list[str]) -> Path:
-    log = tmp_path / "worker_wor-99.log"
-    log.write_text("\n".join(lines) + "\n", encoding="utf-8")
-    return log
-
-
-def test_parse_worker_usage_success(tmp_path: Path) -> None:
-    """Result event provides token snapshot; no compact_boundary → 0 (WOR-357).
-
-    The pre-WOR-357 implementation read context_compactions from the result
-    event's top-level field — that field is never populated by Claude Code.
-    The test now reflects the new behaviour: compactions counted only from
-    type=system, subtype=compact_boundary events.
-    """
-    result_line = json.dumps(
-        {
-            "type": "result",
-            "subtype": "success",
-            "usage": {
-                "input_tokens": 1000,
-                "output_tokens": 200,
-                "cache_read_input_tokens": 0,
-            },
-            "context_compactions": 3,  # ignored under WOR-357 semantics
-        }
-    )
-    log = _write_log(tmp_path, ['{"type":"other","x":1}', result_line])
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok == 1000
-    assert output_tok == 200
-    assert compactions == 0  # was 3 under the old (broken) semantics
-
-
-def test_parse_worker_usage_no_context_compactions(tmp_path: Path) -> None:
-    """Parseable log with no compact_boundary events returns 0 (WOR-357)."""
-    result_line = json.dumps(
-        {"type": "result", "usage": {"input_tokens": 500, "output_tokens": 50}}
-    )
-    log = _write_log(tmp_path, [result_line])
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok == 500
-    assert output_tok == 50
-    assert compactions == 0  # was None under the old semantics
-
-
-def test_parse_worker_usage_missing_log(tmp_path: Path) -> None:
-    """Missing log returns None for all three fields (unchanged)."""
-    log = tmp_path / "no_such_file.log"
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok is None
-    assert output_tok is None
-    assert compactions is None
-
-
-def test_parse_worker_usage_no_result_line(tmp_path: Path) -> None:
-    """Parseable log without usable usage data returns 0 compactions (WOR-357)."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps({"type": "tool_use", "name": "Bash"}),
-            json.dumps({"type": "assistant", "content": "hello"}),
-        ],
-    )
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok is None
-    assert output_tok is None
-    assert compactions == 0  # log was parseable; just no compactions
-
-
-def test_parse_worker_usage_malformed_json(tmp_path: Path) -> None:
-    """Fully unparseable log returns None for all three fields (unchanged)."""
-    log = tmp_path / "worker.log"
-    log.write_text("not json at all\n{broken\n", encoding="utf-8")
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok is None
-    assert output_tok is None
-    # No JSON parseable at all → 0 events seen, but the file IS open-able,
-    # so we get the parseable-but-empty path.
-    assert compactions == 0
-
-
-# ---------------------------------------------------------------------------
-# WOR-357 — compact_boundary system event counting
-# ---------------------------------------------------------------------------
-
-
-def test_parse_worker_usage_one_compact_boundary(tmp_path: Path) -> None:
-    """A single compact_boundary system event yields context_compactions=1."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps(
-                {
-                    "type": "system",
-                    "subtype": "compact_boundary",
-                    "compact_metadata": {
-                        "trigger": "auto",
-                        "pre_tokens": 135486,
-                        "post_tokens": 3348,
-                        "duration_ms": 88463,
-                    },
-                }
-            ),
-            json.dumps(
-                {
-                    "type": "result",
-                    "usage": {"input_tokens": 100, "output_tokens": 50},
-                }
-            ),
-        ],
-    )
-    _, _, compactions, _ = _parse_worker_usage(log)
-    assert compactions == 1
-
-
-def test_parse_worker_usage_multiple_compact_boundaries(tmp_path: Path) -> None:
-    """Three compact_boundary events sum to context_compactions=3."""
-    boundary = json.dumps(
-        {
-            "type": "system",
-            "subtype": "compact_boundary",
-            "compact_metadata": {"trigger": "auto"},
-        }
-    )
-    log = _write_log(
-        tmp_path,
-        [
-            boundary,
-            json.dumps({"type": "assistant", "message": {"id": "a1"}}),
-            boundary,
-            json.dumps({"type": "assistant", "message": {"id": "a2"}}),
-            boundary,
-            json.dumps(
-                {
-                    "type": "result",
-                    "usage": {"input_tokens": 100, "output_tokens": 10},
-                }
-            ),
-        ],
-    )
-    _, _, compactions, _ = _parse_worker_usage(log)
-    assert compactions == 3
-
-
-def test_parse_worker_usage_compact_duration_summed(tmp_path: Path) -> None:
-    """WOR-358: 4th tuple element sums compact_metadata.duration_ms."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps(
-                {
-                    "type": "system",
-                    "subtype": "compact_boundary",
-                    "compact_metadata": {"duration_ms": 50000},
-                }
-            ),
-            json.dumps(
-                {
-                    "type": "system",
-                    "subtype": "compact_boundary",
-                    "compact_metadata": {"duration_ms": 38463},
-                }
-            ),
-            json.dumps(
-                {
-                    "type": "result",
-                    "usage": {"input_tokens": 100, "output_tokens": 5},
-                }
-            ),
-        ],
-    )
-    _, _, compactions, compact_dur = _parse_worker_usage(log)
-    assert compactions == 2
-    assert compact_dur == 88463
-
-
-def test_parse_worker_usage_compact_duration_zero_when_no_compactions(
-    tmp_path: Path,
-) -> None:
-    """No compact_boundary events → compact_duration_ms is 0, not None."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps(
-                {"type": "result", "usage": {"input_tokens": 100, "output_tokens": 5}}
-            ),
-        ],
-    )
-    _, _, compactions, compact_dur = _parse_worker_usage(log)
-    assert compactions == 0
-    assert compact_dur == 0
-
-
-def test_parse_worker_usage_compact_duration_missing_metadata(tmp_path: Path) -> None:
-    """compact_boundary event without duration_ms → counts as compaction but adds 0."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps({"type": "system", "subtype": "compact_boundary"}),
-            json.dumps(
-                {
-                    "type": "system",
-                    "subtype": "compact_boundary",
-                    "compact_metadata": {"trigger": "auto"},  # no duration_ms key
-                }
-            ),
-            json.dumps(
-                {"type": "result", "usage": {"input_tokens": 100, "output_tokens": 5}}
-            ),
-        ],
-    )
-    _, _, compactions, compact_dur = _parse_worker_usage(log)
-    assert compactions == 2
-    assert compact_dur == 0
-
-
-def test_parse_worker_usage_other_system_subtypes_ignored(tmp_path: Path) -> None:
-    """system events with non-compact_boundary subtypes do not increment."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps({"type": "system", "subtype": "init"}),
-            json.dumps({"type": "system", "subtype": "task_started"}),
-            json.dumps({"type": "system", "subtype": "api_retry"}),
-            json.dumps({"type": "system", "subtype": "task_notification"}),
-            json.dumps(
-                {
-                    "type": "result",
-                    "usage": {"input_tokens": 100, "output_tokens": 5},
-                }
-            ),
-        ],
-    )
-    _, _, compactions, _ = _parse_worker_usage(log)
-    assert compactions == 0
-
-
-def test_parse_worker_usage_compact_boundary_with_assistant_usage(
-    tmp_path: Path,
-) -> None:
-    """Cumulative assistant-token sum AND compaction count both reported."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps(
-                {
-                    "type": "assistant",
-                    "message": {
-                        "id": "a1",
-                        "usage": {"input_tokens": 1000, "output_tokens": 100},
-                    },
-                }
-            ),
-            json.dumps({"type": "system", "subtype": "compact_boundary"}),
-            json.dumps(
-                {
-                    "type": "assistant",
-                    "message": {
-                        "id": "a2",
-                        "usage": {"input_tokens": 500, "output_tokens": 50},
-                    },
-                }
-            ),
-            json.dumps(
-                {
-                    "type": "result",
-                    "usage": {"input_tokens": 1500, "output_tokens": 150},
-                }
-            ),
-        ],
-    )
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok == 1500  # 1000 + 500 (sum across assistant events)
-    assert output_tok == 150  # 100 + 50
-    assert compactions == 1
-
-
-def test_parse_worker_usage_cumulative_output_tokens(tmp_path: Path) -> None:
-    """output_tokens summed across every type=assistant event."""
-    assistant = [
-        json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "id": "a1",
-                    "usage": {
-                        "input_tokens": 10000,
-                        "output_tokens": 100,
-                    },
-                },
-            }
-        ),
-        json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "id": "a2",
-                    "usage": {
-                        "input_tokens": 10000,
-                        "output_tokens": 200,
-                    },
-                },
-            }
-        ),
-        json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "id": "a3",
-                    "usage": {
-                        "input_tokens": 10000,
-                        "output_tokens": 300,
-                    },
-                },
-            }
-        ),
-    ]
-    result_line = json.dumps(
-        {
-            "type": "result",
-            "usage": {
-                "input_tokens": 40000,
-                "output_tokens": 707,
-            },
-            "context_compactions": 5,
-        }
-    )
-    log = tmp_path / "worker.log"
-    log.write_text("\n".join(assistant + [result_line]) + "\n", encoding="utf-8")
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert output_tok == 600  # 100+200+300
-    assert compactions == 0  # WOR-357: result.context_compactions field is ignored
-
-
-def test_parse_worker_usage_cumulative_input_tokens(tmp_path: Path) -> None:
-    """input_tokens summed across every type=assistant event."""
-    assistant = [
-        json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "id": "a1",
-                    "usage": {
-                        "input_tokens": 8000,
-                        "output_tokens": 50,
-                    },
-                },
-            }
-        ),
-        json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "id": "a2",
-                    "usage": {
-                        "input_tokens": 12000,
-                        "output_tokens": 60,
-                    },
-                },
-            }
-        ),
-        json.dumps(
-            {
-                "type": "assistant",
-                "message": {
-                    "id": "a3",
-                    "usage": {
-                        "input_tokens": 15000,
-                        "output_tokens": 70,
-                    },
-                },
-            }
-        ),
-    ]
-    result_line = json.dumps(
-        {
-            "type": "result",
-            "usage": {
-                "input_tokens": 50000,
-                "output_tokens": 200,
-            },
-            "context_compactions": 2,
-        }
-    )
-    log = tmp_path / "worker.log"
-    log.write_text("\n".join(assistant + [result_line]) + "\n", encoding="utf-8")
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok == 35000  # 8000+12000+15000
-
-
-def test_parse_worker_usage_mixed_valid_invalid_lines(tmp_path: Path) -> None:
-    result_line = json.dumps(
-        {
-            "type": "result",
-            "usage": {"input_tokens": 300, "output_tokens": 100},
-            "context_compactions": 1,
-        }
-    )
-    log = tmp_path / "worker.log"
-    log.write_text("garbage line\n" + result_line + "\n", encoding="utf-8")
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok == 300
-    assert output_tok == 100
-    assert compactions == 0  # WOR-357: result.context_compactions field is ignored
-
-
-def test_parse_worker_usage_returns_first_result_line(tmp_path: Path) -> None:
-    first = json.dumps(
-        {"type": "result", "usage": {"input_tokens": 10, "output_tokens": 5}}
-    )
-    second = json.dumps(
-        {"type": "result", "usage": {"input_tokens": 999, "output_tokens": 999}}
-    )
-    log = _write_log(tmp_path, [first, second])
-    input_tok, output_tok, _, _ = _parse_worker_usage(log)
-    # No assistant events — fallback uses the last result event's snapshot.
-    assert input_tok == 999
-    assert output_tok == 999
-
-
-def test_parse_worker_usage_empty_file(tmp_path: Path) -> None:
-    """Empty file is open-able and parseable → (None, None, 0)."""
-    log = tmp_path / "empty.log"
-    log.write_text("", encoding="utf-8")
-    input_tok, output_tok, compactions, _ = _parse_worker_usage(log)
-    assert input_tok is None
-    assert output_tok is None
-    assert compactions == 0  # WOR-357: parseable path returns 0, not None
-
-
-# ---------------------------------------------------------------------------
-# WOR-360 — _parse_worker_api_retries
-# ---------------------------------------------------------------------------
-
-
-def test_parse_worker_api_retries_zero(tmp_path: Path) -> None:
-    """Log with no api_retry events returns 0."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps({"type": "system", "subtype": "init"}),
-            json.dumps(
-                {"type": "result", "usage": {"input_tokens": 100, "output_tokens": 5}}
-            ),
-        ],
-    )
-    assert _parse_worker_api_retries(log) == 0
-
-
-def test_parse_worker_api_retries_counts_5(tmp_path: Path) -> None:
-    """5 api_retry events returns 5."""
-    retry = json.dumps({"type": "system", "subtype": "api_retry"})
-    log = _write_log(
-        tmp_path,
-        [
-            retry,
-            retry,
-            retry,
-            retry,
-            retry,
-            json.dumps(
-                {"type": "result", "usage": {"input_tokens": 100, "output_tokens": 5}}
-            ),
-        ],
-    )
-    assert _parse_worker_api_retries(log) == 5
-
-
-def test_parse_worker_api_retries_other_subtypes_ignored(tmp_path: Path) -> None:
-    """Other system subtypes (init, compact_boundary, task_started) ignored."""
-    log = _write_log(
-        tmp_path,
-        [
-            json.dumps({"type": "system", "subtype": "init"}),
-            json.dumps({"type": "system", "subtype": "compact_boundary"}),
-            json.dumps({"type": "system", "subtype": "task_started"}),
-            json.dumps({"type": "system", "subtype": "task_notification"}),
-            json.dumps({"type": "system", "subtype": "api_retry"}),
-        ],
-    )
-    assert _parse_worker_api_retries(log) == 1
-
-
-def test_parse_worker_api_retries_missing_log(tmp_path: Path) -> None:
-    """Missing log returns None (cannot read)."""
-    assert _parse_worker_api_retries(tmp_path / "no_such_file.log") is None
-
-
-# ---------------------------------------------------------------------------
-# WOR-364 — _parse_worker_subagent_spawns
-# ---------------------------------------------------------------------------
-
-
-def _task_use(name: str = "Task") -> str:
-    return json.dumps(
-        {
-            "type": "assistant",
-            "message": {
-                "id": "a1",
-                "content": [{"type": "tool_use", "name": name, "input": {}}],
-            },
-        }
-    )
-
-
-def test_parse_worker_subagent_spawns_zero(tmp_path: Path) -> None:
-    """Log with no Task tool_use returns 0."""
-    log = _write_log(
-        tmp_path,
-        [
-            _task_use("Read"),
-            _task_use("Edit"),
-            _task_use("Bash"),
-            json.dumps(
-                {"type": "result", "usage": {"input_tokens": 100, "output_tokens": 5}}
-            ),
-        ],
-    )
-    assert _parse_worker_subagent_spawns(log) == 0
-
-
-def test_parse_worker_subagent_spawns_counts_3(tmp_path: Path) -> None:
-    """3 Task tool_use events returns 3."""
-    log = _write_log(
-        tmp_path,
-        [
-            _task_use("Task"),
-            _task_use("Read"),
-            _task_use("Task"),
-            _task_use("Bash"),
-            _task_use("Task"),
-            json.dumps(
-                {"type": "result", "usage": {"input_tokens": 100, "output_tokens": 5}}
-            ),
-        ],
-    )
-    assert _parse_worker_subagent_spawns(log) == 3
-
-
-def test_parse_worker_subagent_spawns_other_tools_ignored(tmp_path: Path) -> None:
-    """Read/Edit/Bash/Grep/Write/TodoWrite are not counted."""
-    log = _write_log(
-        tmp_path,
-        [
-            _task_use("Read"),
-            _task_use("Edit"),
-            _task_use("Bash"),
-            _task_use("Grep"),
-            _task_use("Write"),
-            _task_use("TodoWrite"),
-        ],
-    )
-    assert _parse_worker_subagent_spawns(log) == 0
-
-
-def test_parse_worker_subagent_spawns_missing_log(tmp_path: Path) -> None:
-    """Missing log returns None."""
-    assert _parse_worker_subagent_spawns(tmp_path / "no_such_file.log") is None
-
-
-# ---------------------------------------------------------------------------
-# _parse_worker_usage — 3-tuple return (WOR-230)
-# ---------------------------------------------------------------------------
-
-
-def test_parse_worker_usage_returns_separate_tokens(tmp_path: Path) -> None:
-    """input_tokens and output_tokens are returned separately."""
-    result_line = json.dumps(
-        {
-            "type": "result",
-            "usage": {"input_tokens": 12000, "output_tokens": 800},
-        }
-    )
-    log = _write_log(tmp_path, [result_line])
-    input_tok, output_tok, _, _ = _parse_worker_usage(log)
-    assert input_tok == 12000
-    assert output_tok == 800
-
-
-def test_parse_worker_usage_missing_input_token_returns_none(
-    tmp_path: Path,
-) -> None:
-    """When input_tokens is absent, all tokens are None."""
-    result_line = json.dumps({"type": "result", "usage": {"output_tokens": 500}})
-    log = _write_log(tmp_path, [result_line])
-    input_tok, output_tok, _, _ = _parse_worker_usage(log)
-    assert input_tok is None
-    assert output_tok is None
-
-
-def test_parse_worker_usage_missing_output_token_returns_none(
-    tmp_path: Path,
-) -> None:
-    """When output_tokens is absent, all tokens are None."""
-    result_line = json.dumps({"type": "result", "usage": {"input_tokens": 3000}})
-    log = _write_log(tmp_path, [result_line])
-    input_tok, output_tok, _, _ = _parse_worker_usage(log)
-    assert input_tok is None
-    assert output_tok is None
-
-
-# ---------------------------------------------------------------------------
-# WOR-384: vLLM-direct sentinel — assistant.usage.output_tokens always 0,
-# fall back to content-block character estimate (~4 chars/token).
-# ---------------------------------------------------------------------------
-
-
-def _vllm_direct_assistant(input_tokens: int, content: list[dict]) -> str:
-    """Build an assistant event matching the post-WOR-368 vLLM-direct shape:
-    usage.output_tokens always 0 even with non-empty content."""
-    return json.dumps(
-        {
-            "type": "assistant",
-            "message": {
-                "id": "chatcmpl-x",
-                "model": "qwen3-coder",
-                "content": content,
-                "usage": {
-                    "input_tokens": input_tokens,
-                    "output_tokens": 0,
-                },
-            },
-        }
-    )
-
-
-def test_parse_worker_usage_vllm_direct_falls_back_to_content_estimate(
-    tmp_path: Path,
-) -> None:
-    """vLLM-direct emits output_tokens=0; parser must estimate from content chars.
-
-    Pre-WOR-384 every wave-1 ticket recorded local_output_tokens=0 even though
-    the workers shipped real PRs with thousands of generated tokens.
-    """
-    # 4000 chars of thinking + 400 chars of text = 4400 chars → ~1100 tokens
-    thinking_text = "deep thinking " * 250  # 250 * 16 = ~4000 chars
-    text_text = "result text " * 30  # ~360 chars
-    log = _write_log(
-        tmp_path,
-        [
-            _vllm_direct_assistant(
-                50_000,
-                [
-                    {"type": "thinking", "thinking": thinking_text},
-                    {"type": "text", "text": text_text},
-                ],
-            ),
-        ],
-    )
-    input_tok, output_tok, _, _ = _parse_worker_usage(log)
-    assert input_tok == 50_000
-    assert output_tok is not None
-    # Estimate is content_chars // 4. Allow a generous range to keep the test
-    # resilient to small chars/token-ratio tuning later.
-    assert 800 <= output_tok <= 1500, f"got {output_tok}"
-
-
-def test_parse_worker_usage_vllm_direct_includes_tool_use_input(
-    tmp_path: Path,
-) -> None:
-    """Tool-use blocks contribute their JSON-serialized input to the estimate."""
-    log = _write_log(
-        tmp_path,
-        [
-            _vllm_direct_assistant(
-                10_000,
-                [
-                    {
-                        "type": "tool_use",
-                        "id": "t1",
-                        "name": "Bash",
-                        "input": {"command": "x" * 200},
-                    },
-                ],
-            ),
-        ],
-    )
-    input_tok, output_tok, _, _ = _parse_worker_usage(log)
-    assert input_tok == 10_000
-    assert output_tok is not None and output_tok > 0
-
-
-def test_parse_worker_usage_real_anthropic_shape_unaffected(
-    tmp_path: Path,
-) -> None:
-    """When output_tokens is reported truthfully, parser sums them as before.
-
-    Ensures the WOR-384 sentinel doesn't fire on Anthropic API or any future
-    vLLM that fills output_tokens.
-    """
-    line1 = json.dumps(
-        {
-            "type": "assistant",
-            "message": {
-                "id": "msg_1",
-                "content": [{"type": "text", "text": "hello"}],
-                "usage": {"input_tokens": 100, "output_tokens": 25},
-            },
-        }
-    )
-    line2 = json.dumps(
-        {
-            "type": "assistant",
-            "message": {
-                "id": "msg_2",
-                "content": [{"type": "text", "text": "world"}],
-                "usage": {"input_tokens": 110, "output_tokens": 30},
-            },
-        }
-    )
-    log = _write_log(tmp_path, [line1, line2])
-    input_tok, output_tok, _, _ = _parse_worker_usage(log)
-    assert input_tok == 210
-    assert output_tok == 55  # not estimated — actual sum
-
-
-def test_parse_worker_usage_vllm_direct_no_content_no_estimate(
-    tmp_path: Path,
-) -> None:
-    """If output_tokens=0 AND content is empty, no estimate; result is 0."""
-    log = _write_log(
-        tmp_path,
-        [_vllm_direct_assistant(50_000, [])],
-    )
-    input_tok, output_tok, _, _ = _parse_worker_usage(log)
-    assert input_tok == 50_000
-    assert output_tok == 0  # no content to estimate from
 
 
 # ---------------------------------------------------------------------------
